requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cx_Oracle==8.3.0
dnspython==2.7.0
exceptiongroup==1.3.0
fastmcp==2.3.4
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
markdown-it-py==3.0.0
mcp==1.8.1
-e git+https://github.com/alsqja/mcp_claude.git@bb0f839b97ca8400dea5617d6545c91fadf09461#egg=mcp_project
mdurl==0.1.2
mysql-connector-python==9.3.0
openapi-pydantic==0.5.1
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pymongo==4.13.0
python-dotenv==1.1.0
python-multipart==0.0.20
redis==6.1.0
requests==2.32.3
rich==14.0.0
shellingham==1.5.4
sniffio==1.3.1
sse-starlette==2.2.1
starlette==0.46.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uv==0.6.14
uvicorn==0.34.1
websockets==15.0.1
